multiuser login  and signup

+++ routes/data.route.js
@@ -7,8 +7,13 @@ const BugRouter = express.Router();
 
 
 BugRouter.get("/data", async(req,res) =>{
-    const Bug = await BugModel.find()
-    res.send(Bug)
+      try{
+          const Bug = await BugModel.find({userId:req.body.userId})
+          res.send(Bug)
+      }catch(err){
+        console.log(err)
+        res.send("error in get bug data")
+      }
 })
 
 
@@ -30,18 +35,42 @@ BugRouter.post("/create", async(req,res) =>{
 
 BugRouter.patch("/update/:dataId", async(req,res) =>{
     const dataId = req.params.dataId
+    const userId=req.body.userId
     const payload = req.body;
+      try{
+        const bugdata= await BugModel.findOne({_id:dataId})  
+          if(userId!==bugdata.userId){
+            res.send("you are not authnticated")
+          }else{
+            await BugModel.findByIdAndUpdate({_id:dataId},payload)
+            res.send({"msg":"update data created sucessfully"})
+          }
+      }catch(err){
+        console.log(err)
+        res.send({"msg":"item updated successfully"})
+      }
 
-        await BugModel.findByIdAndUpdate({_id:dataId},payload)
-        res.send({"msg":"update data created sucessfully"})
+       
 })
 // ==============
 
 
 BugRouter.delete("/delete/:dataId", async(req,res) =>{
     const dataId = req.params.dataId
-     await BugModel.findByIdAndDelete({_id:dataId})
-     res.send({"msg":"Delete data sucessfully"})
+    const userId=req.body.userId
+        try{
+           const bugdata= await BugModel.findOne({_id:dataId})  
+             if(userId !==bugdata.userId){
+                res.send("you are not authroised")
+             }else{
+                await BugModel.findByIdAndDelete({_id:dataId})  
+                res.send({"msg":"item deleted succesfully"})
+             }
+        }  catch(err){
+            console.log(err)
+            res.send("some error in delted data" )
+        }
+    
 })
